[Rust]: Finish `tw_any_coin` tests refactoring
